Split into separate projects.
We can't use a workspace here, otherwise cargo will have trouble with the conflicting features for some of the crates. For example the use of different MCU selection features in embassy-stm32 will result in conflicts. We still can open the whole repository in vscode/zed by telling rust-analyzer what the subprojects are.